Extract by_groups into class
 * Allows right fetching to be called once per call to by_groups
